% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthosData.R
\docType{package}
\name{orthosData-package}
\alias{orthosData}
\alias{orthosData-package}
\title{orthosData - Data for the orthos package}
\description{
\pkg{orthosData} is the companion ExperimentData package to the  `orthos` R package for mechanistic studies
using differential gene expression experiments. It provides functions for retrieval from ExperimentHub
and local caching of the models and datasets used internally in `orthos`.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fmicompbio/orthosData}
  \item Report bugs at \url{https://github.com/fmicompbio/orthosData/issues}
}

}
\author{
\strong{Maintainer}: Panagiotis Papasaikas \email{panagiotis.papasaikas@fmi.ch} (\href{https://orcid.org/0000-0002-1640-7636}{ORCID})

Authors:
\itemize{
  \item Charlotte Soneson \email{charlotte.soneson@fmi.ch} (\href{https://orcid.org/0000-0003-3833-2169}{ORCID})
  \item Michael Stadler \email{michael.stadler@fmi.ch} (\href{https://orcid.org/0000-0002-2269-4934}{ORCID})
}

Other contributors:
\itemize{
  \item Friedrich Miescher Institute for Biomedical Research [copyright holder]
}

}
